Add a way to skip CI
Until we get https://github.com/taskcluster/taskcluster/issues/5311
fixed, this is a quick hack to avoid wasting CI resources when unneeded.
Unlike a proper CI skip, it will still run the decision task but it will
exit immediately. This means that CI will be marked green though which
is unfortunate but we have to compromise.