Allow changing the "nightly" channel used by packages for testing purposes
Right now we want to test a huge pahkat change but can't take the risk
of releasing it to the nightly channel. Creating a new channel is easy
enough but needs adaptation on the CI side. This changes all tasks to
use a global variable to define what nightly channel they should be
using to install their dependencies and where they should be published.