Oh man if this makes the branch work
Theory is that it's not downloading the specified branch of this repo because it's already downloaded the `main` branch earlier so it's in the list of already-downloaded repos. Hopefully specifying branch in the *first* reference to this GHA repo will fix the issue. If it does, this needs to be refactored to be less esoteric.